+++ Frontend/lib/features/statistics/components/match_card.dart
@@ -1,18 +1,20 @@
-import 'package:darts_application/features/statistics/controllers/completed_matches_controller.dart';
+import 'package:darts_application/features/statistics/stores/statistics_store.dart';
 import 'package:flutter/material.dart';
 import 'package:darts_application/models/match.dart';
+import 'package:provider/provider.dart';
 
 class MatchCard extends StatelessWidget {
   final MatchModel match;
-  final CompletedMatchesController controller;
 
-  const MatchCard({required this.match, required this.controller, super.key});
+  MatchCard({required this.match, super.key});
 
   @override
   Widget build(BuildContext context) {
+    final statisticsStore = Provider.of<StatisticsStore>(context);
     final matchId = int.parse(match.id);
+
     return FutureBuilder<String>(
-      future: controller.getMatchResult(matchId),
+      future: statisticsStore.getMatchResult(matchId),
       builder: (context, snapshot) {
         if (snapshot.connectionState == ConnectionState.waiting) {
           return const Center(child: CircularProgressIndicator());
@@ -27,8 +29,9 @@ class MatchCard extends StatelessWidget {
 
         return Card(
           margin: const EdgeInsets.symmetric(horizontal: 10, vertical: 5),
-          shape:
-              RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(10),
+          ),
           elevation: 2,
           child: Padding(
             padding: const EdgeInsets.all(10.0),
@@ -36,14 +39,16 @@ class MatchCard extends StatelessWidget {
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
                 Text(
-                  '${controller.getPlayerName(match.player1Id)} vs ${controller.getPlayerName(match.player2Id)}',
+                  '${statisticsStore.getPlayerName(match.player1Id)} vs ${statisticsStore.getPlayerName(match.player2Id)}',
                   style: const TextStyle(
-                      fontWeight: FontWeight.bold, fontSize: 18),
+                    fontWeight: FontWeight.bold,
+                    fontSize: 18,
+                  ),
                 ),
                 const SizedBox(height: 5),
                 Text(
                   'Date: ${match.date.toLocal().toIso8601String().split('T').first}',
-                  style: TextStyle(color: Colors.grey[300]),
+                  style: TextStyle(color: Colors.grey[600]),
                 ),
                 const SizedBox(height: 10),
                 Row(
@@ -53,8 +58,8 @@ class MatchCard extends StatelessWidget {
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: [
                         Text(
-                          'Winner: ${controller.getPlayerName(match.winnerId ?? '')}',
-                          style: TextStyle(color: Colors.grey[300]),
+                          'Winner: ${statisticsStore.getPlayerName(match.winnerId ?? '')}',
+                          style: TextStyle(color: Colors.grey[600]),
                         ),
                         const SizedBox(height: 5),
                         Text(finalResult),
@@ -62,7 +67,7 @@ class MatchCard extends StatelessWidget {
                     ),
                     ElevatedButton(
                       onPressed: () =>
-                          controller.navigateToStatistics(match.id),
+                          statisticsStore.navigateToStatistics(match.id),
                       style: ElevatedButton.styleFrom(
                         shape: RoundedRectangleBorder(
                           borderRadius: BorderRadius.circular(10),
